+++ src/main/java/org/betonquest/betonquest/quest/event/point/PointEventFactory.java
@@ -7,13 +7,14 @@ import org.betonquest.betonquest.api.quest.event.EventFactory;
 import org.betonquest.betonquest.config.PluginMessage;
 import org.betonquest.betonquest.data.PlayerDataStorage;
 import org.betonquest.betonquest.instruction.Instruction;
+import org.betonquest.betonquest.instruction.argument.VariableArgument;
 import org.betonquest.betonquest.instruction.variable.VariableNumber;
+import org.betonquest.betonquest.instruction.variable.VariableString;
 import org.betonquest.betonquest.quest.event.IngameNotificationSender;
 import org.betonquest.betonquest.quest.event.NoNotificationSender;
 import org.betonquest.betonquest.quest.event.NotificationLevel;
 import org.betonquest.betonquest.quest.event.NotificationSender;
 import org.betonquest.betonquest.quest.registry.processor.VariableProcessor;
-import org.betonquest.betonquest.util.Utils;
 
 import java.util.Locale;
 
@@ -69,8 +70,7 @@ public class PointEventFactory implements EventFactory {
                 throw new QuestException("Unknown modification action: " + action, e);
             }
         }
-        final String categoryName = instruction.next();
-        final String category = Utils.addPackage(instruction.getPackage(), categoryName);
+        final VariableString category = instruction.get(VariableArgument.STRING_WITH_PACKAGE);
         String number = instruction.next();
         if (!number.isEmpty() && number.charAt(0) == '*') {
             type = Point.MULTIPLY;
@@ -90,6 +90,6 @@ public class PointEventFactory implements EventFactory {
         }
 
         final VariableNumber amount = new VariableNumber(variableProcessor, instruction.getPackage(), number);
-        return new PointEvent(pointSender, categoryName, category, amount, type, dataStorage);
+        return new PointEvent(pointSender, category, amount, type, dataStorage);
     }
 }