+++ CHANGELOG.md
@@ -7,6 +7,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased] - ${maven.build.timestamp}
 ### Added
 - `quest compass` location now allows variables
+- variable support in `point` and `globalpoint` events
 ### Changed
 - message.yml file was deleted and instead the lang folder now contains all translations
 ### Deprecated